storage/requirements.txt
minio==6.0.0
python-dotenv == 0.13.0
nose          ==1.3.7
pytz==2020.1
Flask==1.1.2
Twisted==20.3.0
requests==2.23.0
boto3==1.14.57
